+++ inst/htmlwidgets/rbokeh.js
@@ -5,7 +5,10 @@ HTMLWidgets.widget({
   type: 'output',
 
   initialize: function(el, width, height) {
-    return {}
+    return {
+      modelid: "",
+      elementid: ""
+    }
   },
 
   renderValue: function(el, x, instance) {
@@ -17,16 +20,15 @@ HTMLWidgets.widget({
       x.all_models = JSON.parse(x.all_models);
     }
 
-    console.log(x);
-
-    // set model size in here based on width, height
-
     Bokeh.logger.info("Realizing plot:")
     Bokeh.logger.info(" - modeltype: " + x.modeltype);
     Bokeh.logger.info(" - modelid: " + x.modelid);
     Bokeh.logger.info(" - elementid: " + x.elementid);
 
-    if(x.r_debug == true) {
+    instance.modelid = x.modelid;
+    instance.elementid = x.elementid;
+
+    if(x.debug == true) {
       console.log(x.all_models);
       console.log(JSON.stringify(x.all_models));
     }
@@ -44,7 +46,29 @@ HTMLWidgets.widget({
   },
 
   resize: function(el, width, height, instance) {
-    // (resize code...)
+    // var width = 500;
+    // var height = 500;
+    // var instance = {
+    //   modelid: "80e6788671f9703a699d1f9b98a458d2",
+    //   elementid: "1a247d4cb757568cfdf195a517e1f956"
+    // };
+
+    var box = document.getElementById(instance.elementid).getElementsByTagName("table")[0];
+
+    if(Bokeh.index["c3bcea825e54ec0729458735b959b640"].model.attributes.children) {
+      // it is a gridplot (TODO)
+      // need to get dimensions and set dimensions of all plots
+      // Bokeh.index[instance.modelid].child_views["280e9aeabee465620374997c62fd020b"].canvas._set_dims([300,300])
+    } else {
+      // it's a regular plot
+      var bk_canvas = document.getElementById(instance.elementid).getElementsByClassName("bk-canvas-wrapper")[0];
+      var h_pad = box.clientHeight - bk_canvas.clientHeight;
+      var w_pad = box.clientWidth - bk_canvas.clientWidth;
+
+      // TODO: also shrink font sizes, etc., to a certain degree?
+      Bokeh.index[instance.modelid].canvas._set_dims([width - w_pad,height - h_pad]);
+
+    }
   }
 
 });