+++ NEWS.md
@@ -1,6 +1,13 @@
-Version 0.2
+Version 0.2.1
 ----------------------------------------------------------------------
 
+- add width and height parameters to `grid_plot()`
+- update default padding and text sizes
+- fix sizing so that plots and grid plots on are exact dimensions as specified (resize canvas elements to accommodate for things drawn in the margins)
+- fix bug in generating unique id for each plot
+- allow toolbar to be completely hidden
+- update documentation for `figure()`
+- add sizing policy in rbokeh.js
 - add default nonselection glyph properties for box/lasso select
 - add basic theme support
 - add non-standard evaluation specification of parameters